+++ .gitignore
@@ -1,4 +1,4 @@
-Website-tests/report
+nrt-ftest/Website-tests/report
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
@@ -128,3 +128,7 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+report/
+
+.idea/
