+++ nrt-ftest/main.py
@@ -1,0 +1,4 @@
+import app
+
+if __name__ == '__main__':
+    app.run()
