--- report/TestResults_setUpClass (Tests.home_2019-11-17_08-02-23.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Unittest Results</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
-</head>
-<body>
-    <div class="container">
-        <div class="row">
-            <div class="col-xs-12">
-                <h2 class="text-capitalize">Unittest Results</h2>
-                <p class='attribute'><strong>Start Time: </strong>2019-11-17 08:02:23</p>
-                <p class='attribute'><strong>Duration: </strong>0 ms</p>
-                <p class='attribute'><strong>Summary: </strong>Total: 1, Pass: 0, Error: 1</p>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-xs-12 col-sm-10 col-md-10">
-                <table class='table table-hover table-responsive'>
-                    <thead>
-                        <tr>
-                            <th>setUpClass (Tests.home</th>
-                            <th>Status</th>
-                            <th></th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <tr class='warning'>
-                            <td class="col-xs-10">HomeTest)</td>
-                            <td class="col-xs-1">
-                                <span class="label label-warning" style="display:block;width:40px;">Error</span>
-                            </td>
-                            <td class="col-xs-1">
-                                <button class="btn btn-default btn-xs">View</button>
-                            </td>
-                        </tr>
-                        <tr style="display:none;">
-                            <td class="col-xs-9" colspan="3"><p style="color:maroon;">WebDriverException: Message: 'chromedriver' executable needs to be in PATH. Please see https://sites.google.com/a/chromium.org/chromedriver/home
-</p><p style="color:maroon;">Traceback (most recent call last):
-  File "/home/prat/Documents/epitech/.venv/nrt/lib/python3.8/site-packages/selenium/webdriver/common/service.py", line 72, in start
-    self.process = subprocess.Popen(cmd, env=self.env,
-  File "/usr/lib/python3.8/subprocess.py", line 854, in __init__
-    self._execute_child(args, executable, preexec_fn, close_fds,
-  File "/usr/lib/python3.8/subprocess.py", line 1702, in _execute_child
-    raise child_exception_type(errno_num, err_msg, err_filename)
-FileNotFoundError: [Errno 2] No such file or directory: 'chromedriver'
-
-During handling of the above exception, another exception occurred:
-
-Traceback (most recent call last):
-  File "/home/prat/Documents/epitech/NRT-FTEST/Website-tests/Tests/home.py", line 10, in setUpClass
-    cls.driver = webdriver.Chrome("chromedriver")
-  File "/home/prat/Documents/epitech/.venv/nrt/lib/python3.8/site-packages/selenium/webdriver/chrome/webdriver.py", line 73, in __init__
-    self.service.start()
-  File "/home/prat/Documents/epitech/.venv/nrt/lib/python3.8/site-packages/selenium/webdriver/common/service.py", line 81, in start
-    raise WebDriverException(
-selenium.common.exceptions.WebDriverException: Message: 'chromedriver' executable needs to be in PATH. Please see https://sites.google.com/a/chromium.org/chromedriver/home
-
-</p>
-                            </td>
-                        </tr>
-                        <tr>
-                            <td colspan="3">
-                                Total: 1, Pass: 0, Error: 1 -- Duration: 0 ms
-                            </td>
-                        </tr>
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    </div>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.2.4/jquery.min.js"></script>
-    <script type="text/javascript">
-        $(document).ready(function(){
-            $('td').on('click', '.btn', function(e){
-                e.preventDefault();
-                e.stopImmediatePropagation();
-                var $this = $(this);
-                var $nextRow = $this.closest('tr').next('tr');
-                $nextRow.slideToggle("fast");
-                $this.text(function(i, text){
-                    if (text === 'View') {
-                        return 'Hide';
-                    } else {
-                        return 'View';
-                    };
-                });
-            });
-        });
-    </script>
-</body>
-</html
